Derive angular frequency from temporal

+++ symplyphysics/definitions/temporal_frequency_is_events_per_time.py
@@ -6,7 +6,7 @@ from symplyphysics import (Dimensionless, units, expr_to_quantity, Quantity, Sym
 ## Frequency is the number of occurrences of a repeating event per unit of time.
 ## It is also referred to as temporal frequency, which emphasizes the contrast to spatial frequency and angular frequency.
 
-# Law: f = N / t
+# Definition: f = N / t
 # Where:
 ## N is number of events per time,
 ## t is time,
@@ -27,7 +27,7 @@ def print() -> str:
 
 @validate_input_symbols(time_=time)
 @validate_output_symbol(temporal_frequency)
-def calculate_frequency(events_: int, time_: Quantity) -> Quantity:
+def calculate_frequency(events_: float, time_: Quantity) -> Quantity:
     solved = solve(definition, temporal_frequency, dict=True)[0][temporal_frequency]
     result_expr = solved.subs({time: time_, events: events_})
     return expr_to_quantity(result_expr)
